Address set captain review feedback
