+++ apps/bot/src/discord/utils/get-self-user.ts
@@ -1,0 +1,5 @@
+import type { Message, User } from 'discord.js';
+
+export function getSelfUser(message: Message): User {
+  return message.author;
+}
