+++ apps/bot/src/domains/character/commands/index.ts
@@ -1,4 +1,4 @@
 import { crewCommand } from './crew.js';
-import { setCaptainCommand } from './setcaptain.js';
+import { setCaptainCommand } from './set-captain.js';
 
 export const characterCommands = [crewCommand, setCaptainCommand];
